Persist uploaded files to data/ and DB
Add persistent storage for uploaded files under a new data/ directory and record their paths in the DB.

Changes summary:
- .gitignore: ignore data/ directory.
- backend/src/app.py: introduce DATA_DIR and _move_to_storage helper (uses shutil via run_in_threadpool), move uploaded/temp files into data/{category}, update endpoints (VMAF, PEAQ, PESQ, WebRTC, IQA) to persist files, update DB records, return storage_path in responses, and use safer existence checks before unlinking. Adjust background VMAF task to use persistent path.
- backend/src/db/repository.py: add get_username_by_session, extend insert/update functions and schemas to accept and persist storage_path/degraded_storage_path/noise_storage_path fields for vmaf/peaq/pesq/iqa results.
- backend/src/db/schema.sql: add storage_path columns for vmaf, peaq, pesq, and iqa tables.
- frontend/lib/IQA/IQA.dart: small type fix converting clamp result to double (.toDouble()).

Rationale: persist uploads for later inspection or reprocessing, associate files with sessions/users, and ensure DB reflects file locations.

+++ frontend/lib/IQA/IQA.dart
@@ -380,7 +380,7 @@ class _IQAPageState extends State<IQAPage> {
                 math.pow(sN, 0.45) *
                 math.pow(sP, 0.35);
 
-    return cdi.clamp(0.0, 100.0);
+    return cdi.clamp(0.0, 100.0).toDouble();
   }
 
   String _cameraScoreLabel(double v) =>
